Release 1.2.7 (#2646)
# 1.2.7

Created: March 31, 2025 4:13 PM

# Changelog

## New Features:

- **Gemini Image Generation**: Added support for generating images
straight from Gemini using the `gemini-2.0-flash-exp-image-generation`
model.

## Improvements:

- **Vertex AI**: Improved use of Vertex AI with Gemini Model class to
closely follow the official Google specification
- **Function Result Caching Improvement:** We now have result caching on
all Agno Toolkits and any custom functions using the `@tool` decorator.
See the docs <HERE>.
- **Async Vector DB and Knowledge Base Improvements**: Various knowledge
bases, readers and vector DBs now have `async-await` support, so it will
be used in `agent.arun` and `agent.aprint_response`. This also means
that `knowledge_base.aload()` is possible which should greatly increase
loading speed in some cases. The following have been converted:
    - Vector DBs:
- `LanceDb` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/vector_dbs/lance_db/async_lance_db.py)
is a cookbook to illustrate how to use it.
- `Milvus` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/vector_dbs/milvus_db/async_milvus_db.py)
is a cookbook to illustrate how to use it.
- `Weaviate` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/vector_dbs/weaviate_db/async_weaviate_db.py)
is a cookbook to illustrate how to use it.
    - Knowledge Bases:
- `JSONKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/json_kb_async.py)
is a cookbook to illustrate how to use it.
        - `PDFKnowledgeBase` →
        - `PDFUrlKnowledgeBase` →
- `CSVKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/csv_kb_async.py)
is a cookbook to illustrate how to use it.
- `CSVUrlKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/csv_url_kb_async.py)
is a cookbook to illustrate how to use it.
- `ArxivKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/arxiv_kb_async.py)
is a cookbook to illustrate how to use it.
- `WebsiteKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/website_kb_async.py)
is a cookbook to illustrate how to use it.
- `YoutubeKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/youtube_kb_async.py)
is a cookbook to illustrate how to use it.
- `TextKnowledgeBase` →
[Here](https://github.com/agno-agi/agno/blob/main/cookbook/agent_concepts/knowledge/text_kb_async.py)
is a cookbook to illustrate how to use it.

## Bug Fixes:

- **Recursive Chunking Infinite Loop**: Fixes an issue with
RecursiveChunking getting stuck in an infinite loop for large documents.

---------

Co-authored-by: kausmeows <shuklakaustubh84@gmail.com>
Co-authored-by: ysolanky <yash@phidata.com>