Use built in LiteLLM functionality to validate the environment (#2852)
## Summary

The current `log_warning("LITELLM_API_KEY not set. Please set the
LITELLM_API_KEY environment variable.")` code is raising what might be
considered a false alarm, since LiteLLM doesn't need that env var to be
set, and litellm will automatically search for the correct env var (e.g.
OPENAI_API_KEY, MISTRAL_API_KEY, etc) based on the model.

This PR lets LiteLLM tell if there's an issue. It will throw an
exception if the configuration is incorrect.

## Type of change

- [ ] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ x] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ x] Code complies with style guidelines
- [ x] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ x] Self-review completed
- [ x] Documentation updated (comments, docstrings)
- [ x] Examples and guides: Relevant cookbook examples have been
included or updated (if applicable)
- [x ] Tested in clean environment
- [x ] Tests added/updated (if applicable)

---

## Additional Notes

We're working on supporting agno in
https://github.com/mozilla-ai/any-agent and ran across this warning,
which was a false alarm because LITELLM_API_KEY being unset was not a
real issue.

---------

Co-authored-by: manu <manuhortet@gmail.com>
Co-authored-by: Kaustubh <shuklakaustubh84@gmail.com>