chore: Release 1.5.7 (#3426)
# Changelog

## New Features:

- **Slack App**: Introducing the `SlackApp` to allow you to create
agents for Slack! The app allows agents to respond to individual
messages or on group chats, and it creates threads to respond to
messages.
- **Visualization Tools**: Added `VisualizationTools` that uses
`matplotlib` to give agents the ability to make graphs.
- **Brave Search Tools:** Introduced a new toolkit for integrating
`BraveSearch` that allows agent to search the web using brave search
api.

## Improvements:

- **Pass filters for traditional RAG:** Properly pass
down `knowledge_filters` even if `self.add_references=True` which is a
case for traditional RAG (diff from Agentic RAG)
- **Add infer param to Mem0:** Added infer as a param to `Mem0Tools`

## Bug Fixes:

- **Searxng tool initialization:** Fixed Searxng tool initialization
error
    
    ```
    AttributeError: 'Searxng' object has no attribute 'include_tools'
    ```
    
    and added comprehensive unit tests.
    
- **Fix for enum as a response model for Gemini:** With 1.5.6, a
[bug](https://discord.com/channels/965734768803192842/965734768803192845/1377999191632121926)
was introduced now allowing enum as a data type for Gemini response
model.
- **OpenAI parsing structured output:** With the changes in the OpenAI
library we don't need to parse the structured output separately.

---------

Co-authored-by: Kaustubh <shuklakaustubh84@gmail.com>