fix: Remove tool caching (#3494)
## Summary
Tool caching breaks teams. Team tools are remade for each request, since
the `message` is passed into the factory. We have to reset the tools
every time.

For agents some of the tools are factories, but users can also update
the `tools` at any time and we will not be aware of this. It is best to
reset it every time.

#3422

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ ] Code complies with style guidelines
- [ ] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ ] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)