[fix] Open AI tools missing api key (#3517)
## Summary

This PR addresses a minor issue in OpenAI Tools by ensuring that the
correct API key is included for each client.

(If applicable, issue number: #\_\_\_\_)

## Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other:

---

## Checklist

- [ ] Code complies with style guidelines
- [ ] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ ] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

This limitation can lead to incorrect API usage. The current code uses
the OPENAI_API_KEY from the environment variables instead of the
provided API key when it is available.

Co-authored-by: Mustafa Esoofally <coolmusta@gmail.com>